Refactor: move functions to a separate module
